crud done

+++ axios_crud/src/components/AddUser.jsx
@@ -4,6 +4,7 @@ import {
   FormControl,
   FormControlLabel,
   FormLabel,
+  Grid,
   Radio,
   RadioGroup,
   TextField,
@@ -17,6 +18,9 @@ import { endpoint, endpointUpdate, url } from "../constant";
 import axios from "axios";
 import { useNavigate } from "react-router-dom";
 import { useLocation } from "react-router-dom";
+import dayjs from "dayjs";
+import Swal from "sweetalert2/dist/sweetalert2.js";
+// import "sweetalert2/src/sweetalert2.scss";
 
 const AddUser = () => {
   const navigate = useNavigate();
@@ -30,10 +34,23 @@ const AddUser = () => {
     part: "",
     DateofBirth: "",
   });
+  const [olddata, setolddata] = useState({
+    // fname: "",
+    // lname: "",
+    // cardNo: "",
+    // address: "",
+    // assembly: "",
+    // part: "",
+    // DateofBirth: "",
+  });
   const [update, setupdate] = useState("");
-  const [date, setdate] = useState(null);
   const { state } = useLocation();
+  const [selectedDate, setSelectedDate] = useState(null);
 
+  const handleDateChange = (date) => {
+    setSelectedDate(date);
+    setdata({ ...data, DateofBirth: date });
+  };
   /* ------------------------- Form input change event ------------------------ */
   const handleInputChange = (e) => {
     setdata({ ...data, [e.target.name]: e.target.value });
@@ -41,30 +58,47 @@ const AddUser = () => {
 
   /* ----------------------- Form data submited function ---------------------- */
   const handleSubmit = () => {
+    /*if click the update button put Api Call */
+
     if (update) {
-      //   console.log(localStorage.getItem("Posts"));
-      axios.put(url + endpointUpdate + update, data).then((res) => {
-        setFormData([...formData, res.data]);
+      /*checks if the key is already in the data array and 
+      returns true (Api is not called)if it exists otherwise returns false and 
+      returns false and returns(Api is called)*/
 
-        const storedData = JSON.parse(localStorage.getItem("Posts")) || [];
-
-        // Find the index of the element with id equal to update data
-        const indexToUpdate = storedData.findIndex(
-          (item) => item.id === update
-        );
-
-        storedData[indexToUpdate] = {
-          ...storedData[indexToUpdate],
-          ...res.data,
-        };
-
-        // Update localStorage with the modified array
-        localStorage.setItem("formData", JSON.stringify(storedData));
+      const changesExist = Object.keys(data).some((key) => {
+        return data[key] !== olddata[key];
       });
+
+      if (changesExist) {
+        // check add update data for changes
+        axios.put(url + endpointUpdate + update, data).then((res) => {
+          setFormData([...formData, res.data]);
+          const storedData = JSON.parse(localStorage.getItem("Posts")) || [];
+
+          // Find the index of the element with id equal to update data
+          const indexToUpdate = storedData.findIndex(
+            (item) => item.id === update
+          );
+          // specific index stored in update index in array
+          storedData[indexToUpdate] = {
+            ...storedData[indexToUpdate],
+            ...res.data,
+          };
+
+          // Update localStorage with the modified array
+          localStorage.setItem("formData", JSON.stringify(storedData));
+        });
+      }
     } else {
       axios.post(url + endpoint, data).then((res) => {
-        console.log(res.data);
         setFormData([...formData, res.data]);
+        Swal.fire({
+          position: "top-end",
+          icon: "success",
+          title: "Your work has been saved",
+          showConfirmButton: false,
+          timer: 1500,
+        });
       });
     }
     navigate("/");
@@ -72,8 +106,20 @@ const AddUser = () => {
 
   useEffect(() => {
     if (state) {
+      setSelectedDate(dayjs(state.userData.DateofBirth)); // date of birth set using dayjs format
       setupdate(state.userData.id); // state through find user id
       setdata({
+        ...data,
+        fname: state.userData.fname,
+        lname: state.userData.lname,
+        cardNo: state.userData.cardNo,
+        assembly: state.userData.assembly,
+        address: state.userData.address,
+        part: state.userData.part,
+        DateofBirth: state.userData.DateofBirth,
+      });
+      setolddata({
+        //
         ...data,
         fname: state.userData.fname,
         lname: state.userData.lname,
@@ -94,17 +140,21 @@ const AddUser = () => {
         </Typography>
 
         <TextField
+          required
+          id="fname"
+          defaultValue="foo"
           placeholder="Enter your First Name"
           label="First Name"
           name="fname"
-          required
           variant="outlined"
           value={data.fname}
           onChange={handleInputChange}
-          sx={{ maxWidth: "14rem", marginRight: 4 }}
+          // error={!!error.fname}
+          // helperText={error.fname && "firstname is required"}
+          sx={{ marginRight: 4 }}
         />
 
-        <TextField
+        {/* <TextField
           placeholder="Enter your Last Name"
           label="Last Name"
           name="lname"
@@ -112,9 +162,9 @@ const AddUser = () => {
           variant="outlined"
           value={data.lname}
           onChange={handleInputChange}
-        />
+        /> */}
       </Box>
-      <Box m="1rem">
+      {/* <Box m="1rem">
         <TextField
           placeholder="Enter your Card No"
           label="Card No"
@@ -130,7 +180,7 @@ const AddUser = () => {
           <RadioGroup
             row
             aria-labelledby="demo-row-radio-buttons-group-label"
-            name="row-radio-buttons-group"
+            name="radiobuttons"
           >
             <FormControlLabel
               value="female"
@@ -138,22 +188,26 @@ const AddUser = () => {
               label="Female"
               onChange={handleInputChange}
             />
-            <FormControlLabel value="male" control={<Radio />} label="Male" />
+            <FormControlLabel
+              value="male"
+              control={<Radio />}
+              label="Male"
+              onChange={handleInputChange}
+            />
           </RadioGroup>
         </FormControl>
       </Box>
 
-      <Box ml="1rem">
-        <LocalizationProvider dateAdapter={AdapterDayjs}>
+      <Box m="1rem">
+        <LocalizationProvider
+          dateAdapter={AdapterDayjs}
+          style={{ margin: "1rem 0", width: "50%" }}
+        >
           <DatePicker
-            // name="DateofBirth"
-            label="DateofBirth"
-            value={data && data.DateofBirth}
-            onChange={(date) => {
-              console.log(setdate(date));
-              // handleInputChange
-            }}
-            // onChange={handleInputChange}
+            label="Date of Birth"
+            name="DateofBirth"
+            value={selectedDate}
+            onChange={handleDateChange}
           />
         </LocalizationProvider>
 
@@ -165,7 +219,6 @@ const AddUser = () => {
           variant="outlined"
           value={data.address}
           onChange={handleInputChange}
-          style={{ marginTop: "1rem" }}
         />
       </Box>
       <Box m="1rem">
@@ -186,7 +239,7 @@ const AddUser = () => {
           value={data.part}
           onChange={handleInputChange}
         />
-      </Box>
+      </Box> */}
       <Box textAlign="center" mb="2rem">
         <Button
           variant="contained"
